update use case tests
