+++ app/src/test/java/ru/vssemikoz/newsfeed/usacases/UpdateStorageUseCaseTest.java
@@ -14,37 +14,45 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
+import io.reactivex.rxjava3.core.Single;
+import ru.vssemikoz.newsfeed.data.NewsRepository;
 import ru.vssemikoz.newsfeed.data.NewsStorage;
 import ru.vssemikoz.newsfeed.models.Category;
 import ru.vssemikoz.newsfeed.models.Filter;
 import ru.vssemikoz.newsfeed.models.NewsFeedParams;
 import ru.vssemikoz.newsfeed.models.NewsItem;
-import ru.vssemikoz.newsfeed.usecases.UpdateNewsItemsUseCase;
+import ru.vssemikoz.newsfeed.usecases.UpdateStorageUseCase;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.atLeastOnce;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class UpdateStorageUseCaseTest {
     @Captor
-    private ArgumentCaptor<NewsItem> itemCaptor;
-    private List<NewsItem> exampleNewsList = new ArrayList<>();
-    private List<NewsItem> emptyNewsList = new ArrayList<>();
+    private ArgumentCaptor<NewsFeedParams> paramsCaptor;
+    @Captor
+    private ArgumentCaptor<List<NewsItem>> newsItemsCaptor;
     private NewsFeedParams paramsExample;
-    private Filter filterExample;
+    private List<NewsItem> exampleNewsList = new ArrayList<>();
     @Mock
     NewsStorage newsStorage;
+    @Mock
+    NewsRepository repository;
     @InjectMocks
-    UpdateNewsItemsUseCase updateNewsItemsUseCase;
+    UpdateStorageUseCase updateStorageUseCase;
+
 
     @Before
     public void init() {
         initLists();
         initParams();
+    }
+
+    private void initParams() {
+        Filter filter = new Filter(Category.ALL, true);
+        paramsExample = new NewsFeedParams(exampleNewsList, filter);
     }
 
     public void initLists() {
@@ -66,32 +74,37 @@ public class UpdateStorageUseCaseTest {
         return new String(array, Charset.forName("UTF-8"));
     }
 
-    private void initParams() {
-        filterExample = new Filter(Category.ALL, true);
-        paramsExample = new NewsFeedParams(exampleNewsList, filterExample);
+    @Test
+    public void verifyGetNewsFilteredIsCalled() {
+        when(repository.getNewsFiltered(any())).thenReturn(Single.just(exampleNewsList));
+        updateStorageUseCase.run(paramsExample);
+        verify(repository).getNewsFiltered(any());
     }
 
     @Test
-    public void verifyUpdateItemIsCalled() {
-        updateNewsItemsUseCase.run(paramsExample);
-        verify(newsStorage, atLeastOnce()).updateItem(any());
+    public void getNewsFilteredShouldContainCategory() {
+        when(repository.getNewsFiltered(any())).thenReturn(Single.just(exampleNewsList));
+        updateStorageUseCase.run(paramsExample);
+        verify(repository).getNewsFiltered(paramsCaptor.capture());
+        NewsFeedParams capturedArgument = paramsCaptor.getValue();
+        assertEquals(capturedArgument, paramsExample);
     }
 
     @Test
-    public void updateItemShouldContainNewsItemsExactlyTimes() {
-        updateNewsItemsUseCase.run(paramsExample);
-        verify(newsStorage, times(paramsExample.getNews().size())).updateItem(itemCaptor.capture());
-        List<NewsItem> capturedArgument = itemCaptor.getAllValues();
+    public void verifyInsertUniqueIsCalled() {
+        when(repository.getNewsFiltered(any())).thenReturn(Single.just(exampleNewsList));
+        updateStorageUseCase.run(any()).test();
+        verify(newsStorage).insertUnique(any());
+    }
+
+    @Test
+    public void insertUniqueShouldContainNewsList() {
+        when(repository.getNewsFiltered(any())).thenReturn(Single.just(exampleNewsList));
+        updateStorageUseCase.run(any()).test();
+        verify(newsStorage).insertUnique(newsItemsCaptor.capture());
+        List<NewsItem> capturedArgument = newsItemsCaptor.getValue();
         assertEquals(capturedArgument, exampleNewsList);
     }
 
-    @Test
-    public void updateItemShouldContainNoItems() {
-        NewsFeedParams paramEmptyList = new NewsFeedParams(emptyNewsList, filterExample);
-        updateNewsItemsUseCase.run(paramEmptyList);
-        verify(newsStorage, never()).updateItem(itemCaptor.capture());
-        List<NewsItem> capturedArgument = itemCaptor.getAllValues();
-        assertEquals(capturedArgument.size(), 0);
-    }
-
 }
+
